% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higrad.R
\name{plot.higrad}
\alias{plot.higrad}
\title{Plot a \code{higrad} Object}
\usage{
\method{plot}{higrad}(x, ...)
}
\arguments{
\item{x}{a fitted object of class \code{higrad}.}

\item{...}{additional graphical parameters.}
}
\description{
Produces a coefficient paths for a fitted \code{higrad} object.
}
