% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higrad.R
\name{print.higrad}
\alias{print.higrad}
\title{Print a \code{higrad} Object}
\usage{
\method{print}{higrad}(x, ...)
}
\arguments{
\item{x}{a fitted object of class \code{higrad}.}

\item{...}{additional print arguments.}
}
\description{
Print the coefficients estimates obtained by HiGrad.
}
